+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.19
 
 - Removed redraw debugging.
+- Created GridWidget (#53)
 
 ## 0.4.18