+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Pushrod Change Log
 
+## 0.4.19
+
+- Removed redraw debugging.
+- Created GridWidget (#53)
+
 ## 0.4.18
 
 - Added `SliderWidget`, added `SliderOrientation` for horizontal or vertical orientation. (#30)